Added project #1: Python Chopsticks AI
This AI uses simple machine learning to improve with each game until it is unbeatable and efficient. It also comes with an option to have the game play against another AI and suppress all print statements, which is useful for training the main AI, by simply commenting out one 1 line of the code.