Add files via upload
This is a simple program designed to find the generating function of a constant-recursive sequence. To give a bit more information, a generating series of a sequence s(n) is a function of the form s(0)+s(1)*x+s(2)*x^2+...
Sometimes, these generating series can be simplified. For instance, the generating series of the Fibonacci numbers is x/(-x^2-x+1). If you want to see this using my code, enter 3 for this type of computation and enter the initial conditions s(0)=0, s(1)=1 and s(n+2)=s(n+1)+s(n), which together uniquely define the sequence. The phrase "constant-recursive" just means that the recurrence relation has only constant coefficients (nothing in terms of n or n^2, for instance, in front of s(n+1)).
Have fun testing the code, and feel free to poke around the comments to see how I implemented it with an OOP-heavy approach. More projects on the way!